Environment#expand_if_needed
Expand environment by a row if a living cell is in the last row
Expand environment by a col if a living cell is in the last col

NOTE: This expansion only occurs at the outer edges (max_row and max_col) of the environment. The environment does not expand in a negative direction simply because living cells are located int he first row or column.